% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaderboard.R
\name{ct_leaderboard}
\alias{ct_leaderboard}
\title{Get Leaderboard}
\usage{
ct_leaderboard(
  ids = NULL,
  count = NULL,
  list = NULL,
  start = NULL,
  end = NULL,
  offset = NULL,
  order = NULL,
  sort = NULL,
  token = NULL
)
}
\arguments{
\item{ids}{A list of CrowdTangle accountIds to retrieve leaderboard data for.
These should be provided comma-separated. This and listId are mutually
exclusive; if both are sent, accountIds will be preferred.}

\item{count}{The number of accounts to return on the leaderboard. Default is
50.}

\item{list}{The list of the leaderboard to retrieve. Use \link{ct_getlists} to
retrieve a overview of your lists.}

\item{start}{Start date to create the leaderboard. Format is
“yyyy-mm-ddThh:mm:ss”. Times have to be provided as UTC times. Default is
one day earlier than endDate.}

\item{end}{The endDate of the leaderboard range. Format is the same as start.
Default is now.}

\item{offset}{The number of rows to offset (generally used for pagination).}

\item{order}{"asc" or "desc".}

\item{sort}{"total_interactions" or "interaction_rate". Default is
"total_interactions".}

\item{token}{token.}
}
\description{
Retrieves leaderboard data for a certain list or set of accounts.
}
\details{
For details see https://github.com/CrowdTangle/API/wiki/Leaderboard.
}
