% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{ct_parse_posts}
\alias{ct_parse_posts}
\title{Parse Data from Posts Endpoint}
\usage{
ct_parse_posts(x, recursive = FALSE)
}
\arguments{
\item{x}{files(s) or directory path}

\item{recursive}{if x is a directory, should the function scan recursively
for more files.}
}
\value{
data.frame (tibble)
}
\description{
Given a file or files from a call to \link{ct_posts}, the function parses
the json format into a (mostly) tidy data.frame (tibble). Duplicated entries
(resulting from, e.g., multiple downloads of posts from the same list) are
automatically removed.
}
