% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{ct_posts}
\alias{ct_posts}
\title{Get Posts}
\usage{
ct_posts(
  accounts = NULL,
  brandedContent = NULL,
  count = 10L,
  list = NULL,
  start = NULL,
  end = NULL,
  includeHistory = NULL,
  language = NULL,
  minInteractions = NULL,
  offset = NULL,
  pageAdminTopCountry = NULL,
  searchTerm = NULL,
  sortBy = NULL,
  verified = NULL,
  parse = FALSE,
  data = "./data",
  token = NULL
)
}
\arguments{
\item{accounts}{a string corresponding to account handles (ie iamcardib) or
platform IDs (ie 1436859892).}

\item{brandedContent}{Limits to or excludes posts that have been marked as
Branded Content, either as Publisher or Marketer.}

\item{count}{The maximum number of posts to return.}

\item{list}{The list of the leaderboard to retrieve. Use \link{ct_getlists}
to retrieve a overview of your lists.}

\item{start}{Start date to create the leaderboard. Format is
“yyyy-mm-ddThh:mm:ss”. Times have to be provided as UTC times. Default is
one day earlier than endDate.}

\item{end}{The endDate of the leaderboard range. Format is the same as start.
Default is now.}

\item{includeHistory}{Includes timestep data for growth of each post
returned.}

\item{language}{2-character code of language.}

\item{minInteractions}{If set, will exclude posts with total interactions
below this threshold.}

\item{offset}{The number of rows to offset (generally used for pagination).}

\item{pageAdminTopCountry}{Not sure.}

\item{searchTerm}{Returns only posts that match this search term. Terms AND
automatically. Separate with commas for OR, use quotes for phrases. E.g.
CrowdTangle API -> AND. CrowdTangle, API -> OR. "CrowdTangle API" -> AND in
that exact order. You can also use traditional Boolean search with this
parameter.}

\item{sortBy}{The method by which to filter and order posts: date,
interaction_rate, overperforming, total_interactions, underperforming}

\item{verified}{Limits to posts where the account has the verified setting
matching the input.}

\item{parse}{Logical. Should data be parsed?}

\item{data}{A directory to store json files in.}

\item{token}{token.}
}
\description{
Retrieves posts from accounts of lists of accounts.
}
\details{
For details see https://github.com/CrowdTangle/API/wiki/Posts.
}
