% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list_post.R
\name{get_list_post}
\alias{get_list_post}
\title{Get List Posts}
\usage{
get_list_post(
  listIds,
  startDate,
  endDate,
  token,
  data_path = NULL,
  searchTerm = NULL,
  language = NULL,
  count = 100
)
}
\arguments{
\item{listIds}{string, list id obtained from CrowdTangle dashboard}

\item{startDate}{string, the earliest date at which a post could be posted. Time zone is UTC. Format is “yyyy-mm-ddThh:mm:ss” or “yyyy-mm-dd” (defaults to time 00:00:00).}

\item{endDate}{string, the latest date at which a post could be posted. Time zone is UTC. Format is “yyyy-mm-ddThh:mm:ss” or “yyyy-mm-dd” (defaults to time 00:00:00).}

\item{token}{string, your CrowdTangle API access token}

\item{data_path}{string, path to store data json}

\item{searchTerm}{string, returns only posts that match this search term.}

\item{language}{string, ISO 639-1 locale code}

\item{count}{integer, the number of posts to return, the maximum is 100.}
}
\value{
a dataframe
}
\description{
Get posts from CrowdTangle list. To use the function you need to first create the list in your CrowdTangle dashboard.
}
\examples{
\dontrun{
get_list_post("1234567",
              "2021-06-22T00:00:00",
              "2021-06-22T00:00:10",
              token)
}
}
