% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{parse_posts}
\alias{parse_posts}
\title{Parse Data from Posts Endpoint}
\usage{
parse_posts(x, recursive = FALSE)
}
\arguments{
\item{x}{files(s) or directory path}

\item{recursive}{if x is a directory, should the function scan recursively for
more files.}
}
\value{
data.frame (tibble)
}
\description{
Parse Data from Posts Endpoint
}
